requirements.txt
pillow==10.0.0
pandas==2.0.3
opencv-python==4.8.0.76